pyproject.toml
[tool.poetry]
    authors=["Eoin Farrell, Yeu-Tong Lau"]
    description=""
    include=["unlearning"]
    license="MIT"
    name="unlearning"
    readme="README.md"
    version="0.0.0"

    [tool.poetry.dependencies]
        circuitsvis=">=1.43.2"
        einops=">=0.7"
        python=">=3.10, <3.12"
        torch=">=2.1.1"
        transformer-lens=">=1.11.0"
    scipy = "^1.14.0"
    ipykernel = "^6.29.4"
    matplotlib = "^3.9.0"
    peft = "^0.11.1"
    bitsandbytes = "^0.43.1"
    sae-lens = {path = "../GemmaSAELens"}
    lm-eval = {path = "/root/lm-evaluation-harness"}
    nnsight = "^0.2.21"

    [tool.poetry.group.dev.dependencies]
        plotly=">=5"
        poethepoet=">=0.24.2"
        pydoclint=">=0.3.8"
        pyright=">=1.1.339"
        pytest=">=7"
        pytest-cov=">=4"
        pytest-timeout=">=2.2.0"
        ruff=">=0.1.4"
        syrupy=">=4.6.0"

    [tool.poetry.group.jupyter.dependencies]
        ipywidgets=">=8.1.1"
        jupyterlab=">=4"

[tool.poe.tasks]
    [tool.poe.tasks.check]
        help="All checks"
        ignore_fail=false
        sequence=["check-lock", "docstring-lint", "format", "lint", "test", "typecheck"]

    [tool.poe.tasks.format]
        cmd="ruff format ."
        help="Format code (with autofix)"

    [tool.poe.tasks.lint]
        cmd="ruff check . --fix"
        help="Lint (with autofix)"

    [tool.poe.tasks.docstring-lint]
        cmd="pydoclint ."
        help="Lint docstrings"

    [tool.poe.tasks.ruff]
        help="  [alias for lint && format]"
        ignore_fail=false
        sequence=["format", "lint"]

    [tool.poe.tasks.check-lock]
        cmd="poetry check"
        help="Check the Poetry lockfile"

    [tool.poe.tasks.snapshot-update]
        cmd="pytest --snapshot-update"
        help="Update test snapshots"

    [tool.poe.tasks.test]
        cmd="pytest"
        help="Run tests"

    [tool.poe.tasks.pytest]
        help="  [alias for test]"
        sequence=["test"]

    [tool.poe.tasks.typecheck]
        cmd="pyright"
        help="Typecheck"

    [tool.poe.tasks.pyright]
        help="  [alias for typecheck]"
        sequence=["typecheck"]

[build-system]
    build-backend="poetry.core.masonry.api"
    requires=["poetry-core"]

[tool.pytest]
    cache_dir=".cache/pytest"
    durations=3
    timeout=60

    [tool.pytest.ini_options]
        addopts=[
            "--doctest-modules",
            "--jaxtyping-packages=src,beartype.beartype",
            "-W ignore::beartype.roar.BeartypeDecorHintPep585DeprecationWarning",
            "-s",
        ]
[tool.ruff]
    cache-dir=".cache/ruff"
    extend-include=["*.ipynb"]
    ignore=[
        "ANN101", # self type annotation (it's inferred)
        "F722",   # Forward annotations check (conflicts with jaxtyping)
        "FA102",  # Annotations support (Python >= 3.9 is fine)
        "INP001", # __init__.py for all packages (Python >= 3.3 is fine)
        "PGH003", # No general type: ignore (not supported with pyright)
        "TCH002", # Type checking imports (conflicts with beartype)
        # Rules that conflict with ruff format
        "COM812",
        "ISC001",
    ]
    ignore-init-module-imports=true
    line-length=100
    select=["ALL"]

    [tool.ruff.lint]
        [tool.ruff.lint.flake8-tidy-imports]
            ban-relative-imports="all"

        [tool.ruff.lint.flake8-annotations]
            mypy-init-return=true

        [tool.ruff.lint.isort]
            force-sort-within-sections=true
            known-third-party=["wandb"]
            lines-after-imports=2

        [tool.ruff.lint.per-file-ignores]
            "**/*.ipynb"=[
                "B018", # Useless variables are used for printing
                "D100", # No file docstring needed
                "E402", # Imports don't need to be at the top
                "I001", # Unsorted imports
                "T201", # Print is fine
            ]
            "**/tests/*"=[
                "ARG001", # Fixtures often have unused arguments
                "PT004",  # Fixtures don't return anything
                "S101",   # Assert is needed in PyTest
                "TCH001", # Don't need to mark type-only imports
            ]

        [tool.ruff.lint.pydocstyle]
            convention="google"

        [tool.ruff.lint.pylint]
            max-args=10

[tool.pyright]
    # All rules apart from base are shown explicitly below
    deprecateTypingAliases=true
    disableBytesTypePromotions=true
    include=["."]
    reportAssertAlwaysTrue=true
    reportConstantRedefinition=true
    reportDeprecated=true
    reportDuplicateImport=true
    reportFunctionMemberAccess=true
    reportGeneralTypeIssues=true
    reportIncompatibleMethodOverride=true
    reportIncompatibleVariableOverride=true
    reportIncompleteStub=true
    reportInconsistentConstructor=true
    reportInvalidStringEscapeSequence=true
    reportInvalidStubStatement=true
    reportInvalidTypeVarUse=true
    reportMatchNotExhaustive=true
    reportMissingParameterType=true
    reportMissingTypeArgument=false
    reportMissingTypeStubs=false
    reportOptionalCall=true
    reportOptionalContextManager=true
    reportOptionalIterable=true
    reportOptionalMemberAccess=true
    reportOptionalOperand=true
    reportOptionalSubscript=true
    reportOverlappingOverload=true
    reportPrivateImportUsage=true
    reportPrivateUsage=true
    reportSelfClsParameterName=true
    reportTypeCommentUsage=true
    reportTypedDictNotRequiredAccess=true
    reportUnboundVariable=true
    reportUnknownArgumentType=false
    reportUnknownLambdaType=true
    reportUnknownMemberType=false
    reportUnknownParameterType=false
    reportUnknownVariableType=false
    reportUnnecessaryCast=true
    reportUnnecessaryComparison=true
    reportUnnecessaryContains=true
    reportUnnecessaryIsInstance=true
    reportUnsupportedDunderAll=true
    reportUntypedBaseClass=true
    reportUntypedClassDecorator=true
    reportUntypedFunctionDecorator=true
    reportUntypedNamedTuple=true
    reportUnusedClass=true
    reportUnusedCoroutine=true
    reportUnusedExpression=true
    reportUnusedFunction=true
    reportUnusedImport=true
    reportUnusedVariable=true
    reportWildcardImportFromLibrary=true
    strictDictionaryInference=true
    strictListInference=true
    strictParameterNoneValue=true
    strictSetInference=true

[tool.pydoclint]
    allow-init-docstring=true
    arg-type-hints-in-docstring=false
    check-return-types=false
    check-yield-types=false
    exclude='\.venv'
    style="google"